chore: reorganize disease query code
