+++ src/views/diseases/toc/index.tsx
@@ -3,7 +3,7 @@ import { Flex, Image, Stack } from '@chakra-ui/react';
 import { useQuery } from '@tanstack/react-query';
 import { DiseasePageProps } from 'src/views/diseases/types';
 import { PageContent } from 'src/components/page-container';
-import { fetchAllDiseasePages } from 'src/utils/api/strapi';
+import { fetchAllDiseasePages } from 'src/views/diseases/helpers';
 import {
   StyleCardLabel,
   StyledCard,
